requirements.txt
PyMuPDF==1.23.8
python-docx==0.8.11
scikit-learn==1.3.2
nltk==3.8.1
Sastrawi==1.0.1
pandas==2.1.3
PyYAML==6.0.1
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.0

# Web Framework dan dependencies
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Werkzeug==3.1.3

# Database dependencies
PyMySQL==1.1.1
mysql-connector-python==9.4.0
python-dotenv==1.1.1
bcrypt==4.3.0
SQLAlchemy==2.0.42

# Additional utilities
Pillow==11.3.0
gunicorn==23.0.0 
